Work on Summary and Actions.
Actions are now working: maybe add key elt to the json in order to be able to have a well formated title.
Summary is now Working but need to refine the views as some of the text supperpose each other.
Also need to implement the click on each item of the summary.
A idea would be to display a graph of the last hour in summary list in the background of the item.

Signed-off-by: romain <romain@localhost.localdomain>